refactor: move delete-confirm-modal to shared components
Moves delete-confirm-modal.svelte from the deployments directory to
src/lib/components/delete-confirmation-modal.svelte so it can be reused
across the app. Updates both delete-deployment-modal and
delete-version-modal to import from the new shared location.
