+++ src/lib/components/delete-confirmation-modal.svelte
